move mapLocations to src and update imports
