+++ index.html
@@ -59,7 +59,7 @@
     "generaleventitem": "/components/GeneralEventItem.tsx",
     "iconcomponents": "/components/IconComponents.tsx",
     "interactivemap": "/components/InteractiveMap.tsx",
-    "mapLocations": "/mapLocations.ts",
+    "mapLocations": "/src/mapLocations.ts",
     "weatherforecast": "/components/WeatherForecast.tsx",
     "weatherService": "/services/weatherService.ts"
   }
